fix: crear una nueva app de PatientFile para refactorizar el modelo
