+++ backend/fisio_find/appointment/exampledata.json
@@ -1,5 +1,14 @@
 [
     {
+        "model": "users.pricing",
+        "pk": 1,
+        "fields": {
+            "name": "Plan Básico",
+            "price": 50,
+            "video_limit": 15
+        }
+    },
+    {
         "model": "users.specialization",
         "pk": 1,
         "fields": {
@@ -53,8 +62,7 @@
             "postal_code": "41960",
             "account_status": "ACTIVE",
             "first_name": "Jorge",
-            "last_name": "García Chaparro",
-            "plan": "blue"
+            "last_name": "García Chaparro"
         }
     },
     {
@@ -148,7 +156,8 @@
             "specializations": [
                 1,
                 2
-            ]
+            ],
+            "plan": 1
         }
     },
     {
@@ -163,7 +172,7 @@
             "dni": "12345672A",
             "phone_number": "600000002",
             "postal_code": "28001",
-            "account_status": "ACTIVE",
+            "account_status": "ACTIVE"
         }
     },
     {
@@ -564,8 +573,22 @@
             },
             "patient": 5,
             "physiotherapist": 1,
-            "status": "booked",
+            "status": "finished",
             "alternatives": null
+        }
+    },
+    {
+        "model": "treatments.treatment",
+        "pk": 1,
+        "fields":{
+            "physiotherapist": 1,
+            "patient": 5,
+            "start_time": "2025-03-12T10:00:00Z",
+            "end_time":"2025-03-12T11:00:00Z",
+            "homework": "Realizar ejercicios de estiramiento durante 15 minutos al día.",
+            "is_active": true,
+            "created_at": "2025-03-12T10:00:00Z",
+            "updated_at": "2025-03-12T10:00:00Z"
         }
     }
 ]
