+++ backend/fisio_find/fisio_find/settings.py
@@ -98,6 +98,7 @@ INSTALLED_APPS += [
     'treatments',
     'gestion_survey',
     'payment',
+    'files',
 ]
 
 INSTALLED_APPS += ['corsheaders', 'django_extensions', 'django_filters']
