+++ backend/fisio_find/fisio_find/urls.py
@@ -22,6 +22,7 @@ urlpatterns = [
     path('api/treatments/', include('treatments.urls')),
     path('api/guest_session/', include('guest_session.urls')),
     path('api/payments/', include('payment.urls')),
+    path('api/cloud/', include('files.urls')),
 ]
 
 urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
