+++ backend/fisio_find/users/serializers.py
@@ -634,85 +634,3 @@ class VideoSerializer(serializers.ModelSerializer):
     def get_file_url(self, obj):
         """Devuelve la URL completa del archivo."""
         return f"{settings.DIGITALOCEAN_ENDPOINT_URL}/{obj.file_key}"
-
-
-class PatientFileSerializer(serializers.ModelSerializer):
-    files = serializers.ListField(
-        child=serializers.FileField(),  # Para manejar múltiples archivos
-        write_only=True
-    )
-    file_urls = serializers.SerializerMethodField()
-
-    class Meta:
-        model = PatientFile
-        fields = ["id", "title", "description", "uploaded_at", "file_key", "file_urls", "files"]
-        extra_kwargs = {
-            "file_key": {"read_only": True},  # La clave del archivo es solo de lectura
-        }
-
-    def validate_files(self, files):
-        """Valida que todos los archivos tengan una extensión permitida."""
-        allowed_extensions = (".pdf", ".jpg", ".jpeg", ".png", ".dicom")
-        for file in files:
-            if not file.name.lower().endswith(allowed_extensions):
-                raise serializers.ValidationError(f"Solo se permiten archivos con extensiones {', '.join(allowed_extensions)}.")
-        return files
-
-    def create(self, validated_data):
-        """Sube los archivos a DigitalOcean Spaces y actualiza la BD en una misma subida."""
-        request = self.context["request"]
-        patient = request.user.patient
-        files = validated_data.pop('files')
-
-        # 🔍 Intentamos recuperar una subida previa con el mismo título
-        patient_file, created = PatientFile.objects.get_or_create(
-            patient=patient,
-            title=validated_data.get("title"),
-            defaults={"description": validated_data.get("description"), "file_key": ""}
-        )
-
-        # Inicializamos la lista de claves de archivos
-        file_keys = patient_file.file_key.split(",") if patient_file.file_key else []
-
-        # 🔗 Conectar a DigitalOcean Spaces
-        s3_client = boto3.client(
-            "s3",
-            aws_access_key_id=settings.DIGITALOCEAN_ACCESS_KEY_ID,
-            aws_secret_access_key=settings.DIGITALOCEAN_SECRET_ACCESS_KEY,
-            endpoint_url=settings.DIGITALOCEAN_ENDPOINT_URL
-        )
-
-        for file in files:
-            file_extension = file.name.split(".")[-1]
-            file_key = f"patient_files/{patient.id}/{uuid.uuid4()}.{file_extension}"
-
-            try:
-                s3_client.upload_fileobj(
-                    file,
-                    settings.DIGITALOCEAN_SPACE_NAME,
-                    file_key,
-                    ExtraArgs={"ACL": "private"},  # Mantener privado
-                )
-                file_keys.append(file_key)  # 🔥 Agregamos el archivo a la lista
-            except Exception as e:
-                raise serializers.ValidationError(f"Error al subir archivo: {str(e)}")
-
-        # 🔄 Guardamos la nueva lista de archivos en el modelo
-        patient_file.file_key = ",".join(file_keys)
-        patient_file.save()
-
-        return patient_file
-    
-    def update(self, instance, validated_data):
-        """Actualiza un archivo en DigitalOcean Spaces y la BD."""
-    
-        # Actualizar otros campos si están en los datos
-        instance.title = validated_data.get("title", instance.title)
-        instance.description = validated_data.get("description", instance.description)
-
-        instance.save()
-        return instance
-
-    def get_file_urls(self, obj):
-        """Devuelve la URL completa del archivo."""
-        return f"{settings.DIGITALOCEAN_ENDPOINT_URL}/{obj.file_key}"
